@@ -27,49 +33,48 @@
   "author": "Maarch",
   "license": "GPL-3.0",
   "dependencies": {
-    "@types/node": "^12.12.6",
+    "@types/node": "^12.20.55",
     "angular2-draggable": "^2.3.2",
-    "core-js": "^3.15.1",
-    "tinymce": "^5.8.2",
-    "tinymce-i18n": "^20.12.25",
-    "zone.js": "~0.11.4",
-    "tslib": "^2.3.1"
+    "core-js": "^3.24.1",
+    "tinymce": "^6.1.2",
+    "tinymce-i18n": "^22.5.2",
+    "tslib": "^2.3.1",
+    "zone.js": "^0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.6",
-    "@angular-eslint/eslint-plugin": "^2.1.1",
-    "@angular/animations": "^13.2.6",
-    "@angular/cdk": "^13.2.6",
-    "@angular/cli": "^13.2.6",
-    "@angular/common": "^13.2.6",
-    "@angular/compiler": "^13.2.6",
-    "@angular/compiler-cli": "^13.2.6",
-    "@angular/core": "^13.2.6",
-    "@angular/forms": "^13.2.6",
-    "@angular/material": "^13.3.0",
-    "@angular/platform-browser": "^13.2.6",
-    "@angular/platform-browser-dynamic": "^13.2.6",
-    "@angular/platform-server": "^13.2.6",
-    "@angular/router": "^13.2.6",
-    "@fortawesome/fontawesome-free": "^5.15.3",
-    "@ngx-translate/core": "^13.0.0",
-    "@ngx-translate/http-loader": "^6.0.0",
-    "@swimlane/ngx-charts": "^17.0.1",
+    "@angular-devkit/build-angular": "^14.1.2",
+    "@angular-eslint/eslint-plugin": "^14.0.2",
+    "@angular/animations": "^14.1.2",
+    "@angular/cdk": "^14.1.2",
+    "@angular/cli": "^14.1.2",
+    "@angular/common": "^14.1.2",
+    "@angular/compiler": "^14.1.2",
+    "@angular/compiler-cli": "^14.1.2",
+    "@angular/core": "^14.1.2",
+    "@angular/forms": "^14.1.2",
+    "@angular/material": "^14.1.2",
+    "@angular/platform-browser": "^14.1.2",
+    "@angular/platform-browser-dynamic": "^14.1.2",
+    "@angular/platform-server": "^14.1.2",
+    "@angular/router": "^14.1.2",
+    "@fortawesome/fontawesome-free": "^6.1.2",
+    "@ngx-translate/core": "^14.0.0",
+    "@ngx-translate/http-loader": "^7.0.0",
+    "@swimlane/ngx-charts": "^20.1.0",
     "@types/datatables.net": "^1.10.19",
     "@types/jquery": "^3.5.5",
     "@types/jquery.nicescroll": "^3.7.1",
     "@types/jstree": "^3.3.40",
-    "@types/tooltipster": "0.0.30",
-    "@typescript-eslint/eslint-plugin": "^4.28.0",
-    "@typescript-eslint/eslint-plugin-tslint": "^4.28.0",
-    "@typescript-eslint/parser": "^4.28.0",
-    "ajv": "^7.2.4",
+    "@types/tooltipster": "^0.0.31",
+    "@typescript-eslint/eslint-plugin": "^5.33.0",
+    "@typescript-eslint/eslint-plugin-tslint": "^5.33.0",
+    "@typescript-eslint/parser": "^5.33.0",
     "bootstrap": "^3.4.1",
     "chosen-js": "^1.8.7",
     "codelyzer": "^6.0.2",
     "cypress": "^6.9.1",
     "cypress-file-upload": "^5.0.8",
-    "eslint": "^7.29.0",
+    "eslint": "^8.21.0",
     "eslint-config-prettier": "^8.2.0",
     "eslint-plugin-import": "^2.22.1",
     "flat-to-nested": "^1.1.1",
@@ -80,21 +85,18 @@
     "jquery.nicescroll": "^3.7.6",
     "jstree-bootstrap-theme": "^1.0.1",
     "ng2-dnd": "^5.0.2",
-    "ng2-pdf-viewer": "^6.4.1",
-    "ngx-color": "^6.2.0",
+    "ng2-pdf-viewer": "^9.1.0",
+    "ngx-color": "^8.0.1",
     "ngx-joyride": "^2.4.0",
     "ngx-papaparse": "^5.0.0",
-    "ngx-pipes": "^2.7.5",
+    "ngx-pipes": "^3.1.0",
     "nodemon": "^2.0.7",
-    "npm-check-updates": "^11.5.10",
-    "pdfjs-dist": "2.5.207",
-    "photoswipe": "^4.1.3",
     "prettier": "^2.3.0",
-    "rxjs": "^6.6.7",
+    "rxjs": "^7.5.6",
     "tooltipster": "^4.2.8",
-    "tslib": "^2.3.1",
+    "tslib": "^2.4.0",
     "tslint": "^6.1.3",
-    "typescript": "^4.5.5",
+    "typescript": "^4.7.4",
     "url-loader": "^4.1.1"
   },
   "repository": {}
