@@ -1,5 +1,5 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { FormBuilder, FormGroup, Validators, ValidatorFn } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators, ValidatorFn } from '@angular/forms';
 import { TranslateService } from '@ngx-translate/core';
 import { DomSanitizer } from '@angular/platform-browser';
 import { NotificationService } from '@service/notification/notification.service';
@@ -17,14 +17,14 @@ declare let tinymce: any;
     providers: [ScanPipe]
 })
 export class ParametersCustomizationComponent implements OnInit, OnDestroy {
-    stepFormGroup: FormGroup;
+    stepFormGroup: UntypedFormGroup;
     readonlyState: boolean = false;
 
     backgroundList: any[] = [];
 
     constructor(
         public translate: TranslateService,
-        private _formBuilder: FormBuilder,
+        private _formBuilder: UntypedFormBuilder,
         private notify: NotificationService,
         private sanitizer: DomSanitizer,
         private scanPipe: ScanPipe,
@@ -128,7 +128,7 @@ export class ParametersCustomizationComponent implements OnInit, OnDestroy {
             table_sizing_mode: 'relative',
             table_resize_bars: false,
             toolbar_sticky: true,
-            toolbar_drawer: 'floating',
+            toolbar_mode: 'floating',
             table_style_by_css: true,
             content_style: 'table td { padding: 1px; vertical-align: top; }',
             forced_root_block : false,
